+++ src/main/java/bikerental/BikeType.java
@@ -6,7 +6,7 @@ public class BikeType {
     private final String brand;
     private final String color;
     private String status;
-    private final BigDecimal costPerHour;
+    private final int costPerHour;
     private final int id;
 
     public int getId() {
@@ -22,7 +22,7 @@ public class BikeType {
                 "Cost per hour: " + costPerHour + "| ";
     }
 
-    public BikeType(String brand, String color, String status, BigDecimal costPerHour, int id) {
+    public BikeType(String brand, String color, String status, int costPerHour, int id) {
         this.brand = brand;
         this.color = color;
         this.status = status;
@@ -30,7 +30,7 @@ public class BikeType {
         this.id = id;
     }
 
-    public BigDecimal getCostPerHour() {
+    public int getCostPerHour() {
         return costPerHour;
     }
